third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_after_target_removed_from_slot_mouse-expected.txt

This is a testharness.js-based test.
[FAIL] Pointer events from mouse received before/after slot removal at pointerdown
  assert_equals: events received expected "pointerover@child,pointerenter@host,pointerenter@parent,pointerenter@slot,pointerenter@child,pointerdown@child,(child-removed),pointerover@parent,pointerup@parent,pointerup@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerleave@parent,pointerleave@host" but got "pointerover@child,pointerenter@host,pointerenter@parent,pointerenter@slot,pointerenter@child,pointerdown@child,(child-removed),pointerout@child,pointerleave@child,pointerover@parent,pointerover@host,pointerenter@host,pointerenter@parent,pointerup@parent,pointerup@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
[FAIL] Pointer events from mouse received before/after slotted-child removal at pointerdown
  assert_equals: events received expected "pointerover@child,pointerenter@host,pointerenter@parent,pointerenter@slot,pointerenter@child,pointerdown@child,(child-removed),pointerleave@slot,pointerover@parent,pointerup@parent,pointerup@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerleave@parent,pointerleave@host" but got "pointerover@child,pointerenter@host,pointerenter@parent,pointerenter@slot,pointerenter@child,pointerdown@child,(child-removed),pointerover@parent,pointerenter@parent,pointerup@parent,pointerup@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
[FAIL] Pointer events from mouse received before/after slot removal at pointerup
  assert_equals: events received expected "pointerover@child,pointerenter@host,pointerenter@parent,pointerenter@slot,pointerenter@child,pointerdown@child,pointerup@child,(child-removed),pointerover@parent,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerleave@parent,pointerleave@host" but got "pointerover@child,pointerenter@host,pointerenter@parent,pointerenter@slot,pointerenter@child,pointerdown@child,pointerup@child,(child-removed),pointerout@child,pointerleave@child,pointerover@parent,pointerover@host,pointerenter@host,pointerenter@parent,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
[FAIL] Pointer events from mouse received before/after slotted-child removal at pointerup
  assert_equals: events received expected "pointerover@child,pointerenter@host,pointerenter@parent,pointerenter@slot,pointerenter@child,pointerdown@child,pointerup@child,(child-removed),pointerleave@slot,pointerover@parent,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerleave@parent,pointerleave@host" but got "pointerover@child,pointerenter@host,pointerenter@parent,pointerenter@slot,pointerenter@child,pointerdown@child,pointerup@child,(child-removed),pointerover@slot,pointerenter@parent,pointerenter@slot,pointerout@slot,pointerleave@slot,pointerover@parent,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
Harness: the test ran to completion.
